Fix SRec S5 field handling
S5 records has 2-byte long address field that contain number of data
records expected in the file, but address is read as a 4-byte long hex
value, which results in checksum becoming part of the address read.
This gives incorrect number of data records and as a result the tools
rejects any file with S5 records.